kobe-bryant-shot-selection/requirements.txt
# local package
-e .
xgboost==1.4.2
pandas==1.0.5
pandasql==0.7.3
sklearn
scipy==1.5.4
numpy==1.19.4
# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1